chore: improve Checkbox exports
